main_backend/pyproject.toml
[project]
name = "main-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.1.6",
    "django-extensions>=3.2.3",
    "djangorestframework>=3.15.2",
    "djangorestframework-simplejwt>=5.4.0",
    "djoser>=2.3.1",
    "nanoid>=2.0.0",
    "dj-database-url>=2.3.0",
    "psycopg2-binary>=2.9.6",
    "gunicorn>=21.2.0",
    "python-dotenv>=1.0.1",
]
